# Retrieves OutputPrintable instead of Output