Simplify the API v1 output endpoint. We now always return a single object called OutputPrintable instead of different object when it is by id or by height.
Add method `get_unspent_output_at_pos`.

Before:

```
[
  {
    "commit": "08392e391abc679f6a66478fab4febb783d11904fc9b32cdeabf442ff6026f4baa",
    "height": 375611,
    "mmr_index": 4102967
  }
]
```

After:
```
[
  {
    "output_type": "Transaction",
    "commit": "08392e391abc679f6a66478fab4febb783d11904fc9b32cdeabf442ff6026f4baa",
    "spent": false,
    "proof": "1e93a916bccf08746147c9310954c720140faab878c73558f9a515ef1db314b57affa22d080fc8c0bbb8efa578089bd38025ac347f423648b71adf19bf93416c0cb98805a9eb9e27403621201fa47c89b4b44c864edc8c9123d8890d6194aeb0e13ad0a99d87b265ef79c7940b6d237c6da0b7cdc87badfca921d0598f33c09a3c9f72f3e909afb75d1e06334ab3f0c8a23374f7101d7bd03906f393cc9b1a4de04b561ca6750bffa4c2fbc94f8c8874e289e908665f3d25fdc8fd1707d90e135c640c590b20f00d57f96251dbe84a1cadad8fd06fc4f1556b0da934c5a096eed9788b96464399ce49aeabedf66f8685e6dacef0df781512313fe67e49c27ab47b9c7ee8c2f8d7810d9ec354113da563999e357d754d25750a22dfc10fe9f5482689aee0ac24e64ca48c9639564e6324160591a9ebb8320e13842856f6022ee8361d15f8f1d618a654f9c7fa012e3ba51e188418fdd89b16ae1fd6390051090497810131de7b15898d811bd35efb3ed9f6629ed56f7c25d4aeb669c7dc76077926ac70a1805d0390c0cd13de626034517772d87c43289d9c8ee665fa62837f9b9fc47a9a9e5aff421c4c7c5c5150645e74855cc437bbd9381fbf504dbf22c8c1296c34df202d921b1ba539c113d49247c698d53f7115ee8a202cfb4d15017b9a027a8f51d57b2582eb976bb60593a17e3534b97711c038256e83d801b0b4100c5312ae559e87782687c53861718dbf6367962fe5fbef77ceed9c799df25e2a4f9a795963d2e86f4c09c3c45da87e98c2b4aaabb47a4457d639371ea10ce7649c6434a825115e6452e7f241ddeb15b76145d8c682c6efe185b1cb0467e2b3964fa149a4c4865de58ad5b18edaa93670040e0e9ab997f56d585b07cd58379edb9dc5773e37ed809441ee28b48635b9aaf97d1456e753be6d822d5ee724059afa67ab1e97",
    "proof_hash": "a51aa12fba6316518368f045de5b4658cea859d6da386900e646d0445c5254bb",
    "block_height": 375611,
    "merkle_proof": null,
    "mmr_index": 4102967
  }
]
```