@DavidBurkett 2019-10-04
This is used under the hood during wallet updates here: https://github.com/mimblewimble/grin-wallet/blob/11481669ef14f2d36ba86631ca6464b79a03098a/libwallet/src/internal/updater.rs#L357

So we may need some performance numbers before making this change, since it will inevitably slow that down.