@quentinlesceller 2019-10-04
For the records with 3800 outputs.
With this PR:
`grin-wallet --floonet -p "" outputs` => 1.1s
`grin-wallet --floonet -p "" info` => 0.4 ms.
Without:
`grin-wallet --floonet -p "" outputs` => 0.5 ms
`grin-wallet --floonet -p "" info` => 0.2 ms.

So there is a strong performance hit here....