@antiochp 2019-10-04
> grin-wallet --floonet -p "" outputs => 500 s

That's "500 ms" not "500 s"?