@antiochp 2019-10-10
Out of the loop for a bit here - this is backward compatible right? We're simply adding fields to the existing json?
Or is there a difference here with `height` vs `block_height`? Is this something we need to support for compatibility?