@quentinlesceller 2019-10-10
Because of the `height`/`block_height` thing this is not backward compatible. That's why there is https://github.com/mimblewimble/grin-wallet/pull/235. But if you think this is not worth it (to introduce such breaking change), we might as well simply leave the v1 as is and keep the improvements for the v2.