@quentinlesceller 2019-10-11
As much as I think this is a good improvement, I don't see any urgent to introduce backward incompatibility in the v1 API right now. I'll keep the improvements for v2. Closing.